下一步：分步测试核心函数，找出性能问题

+++ CPU_BC.cpp
@@ -70,7 +70,7 @@ std::vector<float> BC_cpu_parallel(Graph g, int num_cores) {
 	int *R;
 	int *F;
 	int *C;
-	double *result_cpu;
+	float *result_cpu;
 	std::vector<float> result(g.n,0);
 	int n = g.n;
 	int m = g.m;
@@ -78,13 +78,13 @@ std::vector<float> BC_cpu_parallel(Graph g, int num_cores) {
 	R = (int *) _mm_malloc(sizeof(int) * (n + 1), 64);
 	F = (int *) _mm_malloc(sizeof(int) * (m * 2), 64);
 	C = (int *) _mm_malloc(sizeof(int) * (m * 2), 64);
-	result_cpu = (double *) _mm_malloc(sizeof(double) * n * num_cores, 64);
+	result_cpu = (float *) _mm_malloc(sizeof(float) * n * num_cores, 64);
 
 	std::memcpy(R, g.R, sizeof(int) * (n + 1));
 	std::memcpy(F, g.F, sizeof(int) * (m * 2));
 	std::memcpy(C, g.C, sizeof(int) * (m * 2));
 
-	std::memset(result_cpu, 0, sizeof(double) * n * num_cores);
+	std::memset(result_cpu, 0, sizeof(float) * n * num_cores);
 
 	MIC_Opt_BC(n, m, R, F, C, result_cpu, num_cores);
